Work on rebuilding the condition backend and frontend
Condition backend now keeps the comparison data private as well, so that the presence or absence of a global can be properly managed using the flags. I should do the same with the run-on data.

Removed unused code from the condition function and parameter definitions.

Started working on getting the frontend to use the new backend stuff.