Continued work on conditions
All of the code that references conditions has been altered to work with the new backend design, but we've run into a problem: std::vector doesn't like our new condition backend, and the compiler messages are completely unhelpful. The lack of a default constructor is making the condition struct incompatible with std::vector despite online documentation which says that that should be allowed (at most constraining what vector operations you can perform).

I need to rethink a few things. It might be best to have a (condition_list) struct that keeps track of the owner, and then make the owner field on the condition optional; using a struct for the list means that I can make the list itself require an owner, and thereby ensure that any conditions in the list are properly created with an owner.