Work on condition editing: interdependent parameters done
Fixed the union-decider for the VATSValue condition parameter type, so that it returns the None type instead of nullptr. At some point, we may wish to make it so that all type getters related to conditions return condition_parameter_type&, defaulting to None in cases where we currently return nullptr.

Story manager event members now have lists of valid form types, where applicable.

The condition editing UI now handles interdependent parameters, for both standard condition functions and GetEventData.

The condition parameter UI now properly filters form types for the event form parameter, using the form type information on the event member definitions.

Fixed a crash in ConditionParameterEditor::clear which would occur when a parameter's index was set to 2 or higher. Conditions can only have two standard parameters, but have three event parameters, which is both why the crash occurred when the index was set to 2 and why the index would ever be set to 2.