Started work on a helper function for condition descriptions
Started work on a frontend helper function for getting the description of a condition. We should show these as mouseover text when the condition function drop-down is open, and we should remove descriptions from the backend definitions.

Replaced *some* placeholder labels in the quest dialog with condition lists. We should break those tabs up into sub-widgets anyway, so that it's easier to edit the code for them and easier to work with them in Qt Designer.

Fixed a typo in the name of the GetKeywordDataForAlias condition.