Work on Papyrus: separating sub-Quest data out
Continuing to work on separating out "sub-form" script data, currently just for Quests. Moving this data out of the core Quest::script_data struct and into the Quest substructures that it's actually associated with will hopefully make it easier to manage.

ConditionList now emits its conditionEdited signal when the user attempts to reorder conditions, and when the user adds or deletes conditions. Note that in the former case, we emit the signal even if the move operation is impossible.