Papyrus sub-form separation complete but messy
We've done what we set out to do, but it feels messy. I don't like having to pass data back from the form to the Papyrus root when saving, and I don't like that the form loader has to "claim" it from the Papyrus root.

Within the VMAD subrecord, there's the standard data and then there's form-type-specific data immediately afterward. Technically, we can just have the load/save code for specific forms handle that data when appropriate. Maybe we should.