Work on quest-specific VMAD data
The general-purpose VMAD load/save code no longer handles quest-specific VMAD data; the quest loader is now wholly responsible for that. We'll want to get the other form types working the same way once we actually define them.

Because quests are now wholly responsible for quest-specific VMAD data, we no longer need to pass data to the VMAD save code, so we can (and eventually will) remove script_data_save_parameters. We can also consider the stage and entry values in the log entry fragment stale post-load.

The quest loader can now handle alias script data and log entry script fragments. Save code is also present. Use info code is not.

Filled Quest.cpp to the brim with static-asserts as an improvised to-do list. As a consequence of this, the program no longer builds and none of the changes in this commit can be tested.