Work on Papyrus and Quest data
I've decided to have aliases just use a normal (script_data) instance for their Papyrus data, since the format should absolutely just... allow that. I've also decided not to retain Papyrus data that attaches itself to non-existent aliases. However, the latter decision has created a problem.

Added functions to wholly skip the use info for Papyrus data. We use these for when a quest alias claims to belong to a quest other than the containing quest. I don't know if the game would handle that sensibly, but our form-loading system makes it impossible, so we just emit a warning and skip that data during load and thus also during use info generation.

Typically VMAD comes before all quest alias subrecords (and even if that weren't the case, the game still allows it), so when we're building use info for it, we can't guarantee that there's alias data available to check the VMAD data against. We also can't simply hold the VMAD data for reading later. How, then, do we delay generating the use info until we know what aliases are present?

The solution I've thought of is to make (form_stub_use_info_builder) virtual, so that we can subclass it and override the functions for generating use info, to allow us to hold all such calls and then commit them all at a later time if we so choose. This would allow us to instantiate a builder for each newly-encountered alias, and pass that to the VMAD use info function. I've created a skeletal and entirely incomplete (form_stub_deferred_use_info_builder) definition.

Of course, there are potential performance issues with this design that we could avoid if we did things another way. See Quest.cpp circa line 1017 for a writeup.