Improved performance for form_stub_use_info_builder
Improved performance for the form_stub_use_info_builder by having it default to a fixed-size array for storing entries, with a vector as a fallback. This probably won't have as pronounced an impact on Release since STL containers are generally much cheaper there to begin with, but on Debug, it lowers load times to about 9.1 seconds, actually saving more than a second over baseline.