Use info safety for DIAL, DLBR, SHOU, WRLD
DialogueBranch, Topic, Shout, and Worldspace are now protected against records that incorrectly contain multiple instances of certain subrecords (specifically, subrecords which are meant to only appear once and to refer to another form). We will no longer mishandle this kind of malformed data for these forms when generating use info.