Work on form type use info fixes
Wrote some internal documentation on how precisely use info is built, since it turns out we've been doing it wrong.

Converted the Activator and Faction code from the wrong approach to the right one, BUT I've added static asserts as well because there are some things we need to double-check.