TESModel loader fixes
TESModel no longer retains texture swap entries that have no texture set or no NIF block name; this is consistent with the game.

When the form version is below 0xF, TESModel now uses -1 as the NIF block index for texture swap entries. In-game tests establish that in these cases, the entries will never work; still, it's what the game does, so we're doing it too.