Fixes; experimentation with Form::get_papyrus_data overrides
MSVC/MSBuild chose to include zlib not from here, but from the folder where I was keeping my old first draft of DovahKit. When I renamed that folder, everything broke. I have repaired the project and solution files here so that they use the zlib that's actually part of this solution.

Came up with a nice hack to quickly and easily define form-component accessors on subclasses, without having to remember to type them all out every time. I now know how to create a templated function that returns a member if a class has one, or nullptr otherwise.