Fixed compiler errors for previous commit; program now builds
Benchmarking shows that we're three seconds slower than the previous attempts at safely generating extra-data use info (attempts that I apparently forgot to commit as well, because life just sucks, I guess). How the hell is that possible?!

The previous attempts included having a vector of use info builder pointers, heap-allocating one for each extra-data type we find. That should be creating and destroying vastly more space on the stack than what we're currently doing, with the additional overhead of heap-allocation. The only potential problem I can think of is that we're calling the function to send outbound references for every extra-data form ID slot, regardless of whether it's empty. Is that really the source of the overhead?