Fixes to quest stage editing
Fixed crash in QuestTabStages::_select_stage that could occur if the specified stage ID did not exist in the list.

Fixed oversight that caused us to fail to update the stage list after adding a new stage.

Fixed mistake that caused us to select the wrong adjacent stage after removing a stage.