Save fixes; work on quest stages
Added an error handler for if subrecord::write_length_prefixed_string writes a string that's too long. If this works (testing is pending), it'll write placeholder text and report an error, though it can't abort the form write on its own.

Fixed QUST/DNAM save code.

Fixed the quest stage UI to properly update the state of controls for flags and forms.

The Papyrus fragment editor can now have maximum lengths set for the script and function names. The quest stage UI makes use of this as appropriate.