Work on rebuilding the Object Window's treeview
I've started work on rebuilding the Object Window's treeview, with the goal of adding support for quests' filter strings. As part of these efforts, quests' filter strings have been moved from the loaded-form data to the form stub addenda.

Made it so that cloning a form now clones any applicable parts of the form stub addenda.