Started work on custom parse ops and DovahKitFormDataCache
Started work on a system that'll allow us to pull Object Window-relevant fields without having to fully load all forms, and without having to move those fields to the form stub addenda when they don't belong there. (In fact, we'll be moving QUST/FLTR back into the loaded-form data once this system is fully online).

The (form_stub::do_custom_parse) function can be used to perform custom parsing on all of the records that contribute to a given form.

The DovahKitFormDataCache singleton is going to be responsible for caching Object Window-relevant fields, using custom parse operations when data is initially loaded, and using form-modification signals afterward.