Object Window / DovahKitFormDataCache integration
The Object Window is now fully integrated with the DovahKitFormDataCache. This means that we can now maintain a cache of data and have the Object Window rely on that cached data for showing filter sub-trees and for actually applying filters.

The only issue now is that building the cache incurs about a second of delay, and I don't expect that to scale well as we add 'MODL' parsers for more record types. Can we multi-thread it?