Work on condition list flexible columns
Worked on making the condition list columns auto-size to fill the table, while being able to control how the columns consume the available space (rather than just having the last column stretch). That works; however, it also means that if the user resizes the columns, their changes will be lost the next time the containing widget is resized.