Work on QHeaderViewDKEx resize support
QHeaderViewDKEx allows column resizing, but it's jittery, and resizing a flexible column to be too large can prevent the content from fitting in the view.

QHeaderViewDKEx attempts to fill any space that gets lost to integer rounding during the flex calculations. The particular way we do this may be the source of the jitter we're seeing.