Started work on UI scripting
I'm burned out on the form work, so let's skip ahead a bit just so I can get some momentum up.

Added a QWidget pointer to the Lua wrapper class, and added a utility function for wrapping QWidgets.

Added vectors to the editor script core, for storing scripted windows and orphaned widgets (i.e. any widgets that have been created and not yet appended to a window; any widgets that have been removed from windows).

Added a script-to-main cross-thread task for spawning a new window.

Started work on a wrapper-class for QDialogs.