Work on UI scripting: hello world milestone
The DovahKitScriptVM now only runs the teardown function on the main thread. This is necessary since that function is now responsible for cleaning up Qt widgets, which can only be done on the main thread.

DovahKitScriptVM::mainThreadLoop now properly processes outbound UI tasks.

Switched the order of the loop in DovahKitScriptVM::_script_thread_loop to make sure that all outbound messages (including non-blocking fire-and-forget messages) are processed before we check whether we should keep the VM alive. This is needed so that "open window" tasks from the VM have time to complete and be recognized.

With these changes, we can now run a simple script that pops an empty dialog box, and the engine will only kill the Lua VM after the dialog is closed.