Work on UI support: event fixes
Previously, we fired events to Lua from the main thread, which compromised thread-safety. After realizing this, I've changed things so that the main thread sends a main-to-script message to the script thread, which then fires the event to Lua.

Re-enabled the queued functions handler, and optimized it.

The cobb::lua::isempty function now handles negative indices properly.