Fixed file flags handling; work on form scripting
The (file_loader) class now determines whether the file is light or a master based on both the flags in the header and the file extension, consistent with the game.

When saving a file, the (file_load_order) detects whether we're saving it as a light plug-in based on both the header flag and the desired file extension.

Added member functions to the Lua form class for accessing the list of source files, and the last source file. Currently, the only information available per-file is the filename ,effective master flag, and effective light flag.

The pre-made Count Persistent Refs script now shows the number of forms that are persistent due to the flag, and the number of forms that are persistent due to existing inside of a non-master file.