Improvements on previous commit
We now use a "detail" flag in the file header to track whether a file is forced to "light" or "master" by its file extension.

The file loader now makes a distinction between a file's "desired" path and its "actual" path. If a save operation succeeds but a rename operation fails to put the file where it ought to be, the file's effective "light" and "master" states will be determined by the file's desired path (i.e. whether we meant to save it with a file extension that forces those flags), not its actual path (i.e. the temporary file that we managed to save).

That said, a better approach to this might be to use different file extensions for temporary files (e.g. "esptmp", "esltmp", "esmtmp") and then ditch the desired/actual distinction.