Work on UI scripting: UI (un)locking on demand
Implemented (ui.lock) and (ui.unlock). However, the latter function has exposed an issue.

When the main thread sends a UI event to the script thread, it has to access a thread lock on the information queue. When the script thread is actually processing the contents of that queue, it has the lock. This means that if the main thread tries to send an event while a Lua listener is running (which can happen if ui.unlock() has been called), the main thread will immediately be blocked until the script thread has finished sending the current batch of UI events. This is obviously really bad and wholly defeats the purpose of our threading model.

We need to build an alternate queue for UI events, where the script thread keeps the lock only long enough to move queue items to a local vector, before then releasing the lock and processing the queue items without it.