Work on FormPicker
All FormPickers now rely on a three-level model for listing forms.

The innermost level is a single vector of form stubs, with stubs automatically inserted and removed to match the underlying core data.

The middle level is a model that takes advantage of QAbstractItemModel::fetchMore and a QTimer, to filter the form stubs over multiple ticks. This means that the program won't hang while filtering is taking place.

The outer level is a QSortFilterProxyModel which is responsible for sorting the filtered items. Unfortunately, this does occur all at once, and so hangs the program briefly the first time a massive combobox (i.e. one with thousands of items) is opened.

The benefit of the innermost model is that the FormPickers don't need to handle DovahKitCore form-modification events on an individual basis, which allows the FormPicker internal logic to be simpler. The benefit of the middle model is that it greatly helps to alleviate lag associated with FormPicker population. The outer model is needed for sorting, but should be merged into the middle model (i.e. I should implement sorting over time as well) to avoid lag.

At some point, I may even want to investigate the possibility of just straight-up making an asynchronous/across-time replacement for QSortFilterProxyModel that can work for general cases, including trees. That's not a priority right now, though.