Work on FormPicker
The FormPickerSharedUnderlyingModel now stores constructed "item" structs instead of bare form stub pointers. These structs cache the form type locally as well as the editor ID, so that we don't have to convert the editor IDs to QStrings every single individual time we read them (which, during sort calls in particular, is pretty nasty).

Gave FormPickerIterativeModel the ability to sort its contents over time, and removed the QSortFilterProxyModel.

With the above two changes, FormPicker's lag has been completely eliminated even when compiled in Debug, even when it's set to show all of the form types allowable by GetIsID conditions all at once.

Fixed keyboard navigation within FormPicker. Fixed the bug that caused it not to properly update when data was acquired if it was created before data was acquired. Fixed a bug that sometimes caused it not to know that its form type picker was supposed to be visible. Removed the activate-on-show code, since that isn't needed now that a shared model is used. Fixed bugs involving the subwidgets not being enabled or disabled at appropriate times. The form type picker is now sorted.