Work on the Object Window form table
We no longer rebuild the Object Window form table model in order to filter by form type or filter string. We still currently need to be able to do the filtering at the build step in order to process the "all" option; we'll need to make the treeview contents constexpr, basically, to avoid that.