FormPicker fixes and integration
The cell-editing dialog now uses FormPicker instead of FormsOfTypeCombobox.

Fixes to FormPicker internals to prevent value loss when setting the value while the combobox is still filling.

Modified FormPicker to allow setting a default form ID, which will be used if you attempt to set the FormPicker to "none" when "none" isn't allowed.

Fixed "buddy" widgets in the Shout and Shout Word UI files, which were lost when replacing FormsOfTypeComboboxes with FormPickers.