Work on UI scripting: hyperlinks
The script VM now has helper functions for messing with Qt::Alignment values.

Hyperlinks in scripted text are now clickable and show a hardcoded confirmation dialog. The dialog's text changes depending on whether the link is recognizable as an NXM URL or a NexusMods mod page URL.

Added stub code for a UI HTML permission.

Added an API for controlling whether scripted UI text is selectable.