Work on form scripting: quest and alias wrappers
Made quests and their aliases available to Lua.

The generic "wrapper" class for Lua now supports indexed collections where indices are non-contiguous.

We can potentially improve this if we build a custom metatable for the (quest.aliases_by_id) field, allowing it to use alias IDs as raw keys without needing to stringify them to work with the collection boilerplate.