Work on improving script collection wrappers
Started work on improving how we define script collection wrappers. We now have a "params" struct that allows us to pass C-functions in by name, which will make it easier to pass just the parameters we need for any given collection; if we upgrade the project to C++20, then we'll be able to take advantage of designated initializers to make that even easier.

Additionally, the collection internals have been updated to allow numeric-index-only collections to have member functions. Currently, we allow "insert" and "remove" member functions to be provided. In theory, we could also provide named data members through the same mechanism.