Upgraded the project to C++20
Upgraded the project to C++20. This caused a slew of compiler errors that needed to be fixed; additionally, some actual bugs were found during testing.

In C++20, MSVC no longer allows you to pass a temporary to a function that takes a non-const reference argument. Word on the 'net is that this limitation is the C++ standards committee's attempt to prevent developers from shooting themselves in the foot... by blasting off their lower leg with a shotgun, if we're to continue the metaphor. This necessitated some changes to some of the internals, primarily form_load::log_load_warning.

The loaded_form_ptr class definition needed to be moved beneath form_stub, which makes sense since it refers directly to content in form_stub. In retrospect I have no idea how the hell it managed to compile before. This caused some issues with one of form_stub's functions, which itself used functionality in loaded_form_ptr, but fortunately the "declare the function in the class body and define it elsewhere" trick can be done entirely within a single header file, and not just split across H and CPP files.

Every case where QByteArray is stuffed into a std::string now needs us to access QByteArray's data as a char pointer, to avoid ambiguity.

Every constexpr char* value needed to be converted to a constexpr const char* value. Seems a bit redundant, but okay. This is responsible for the bulk of the changes to the Lua VM files.

The check for std::endian support in my endianness helper was incorrect, and has been fixed.

MSVC under C++20 no longer allows fields to be named "default" despite the keyword having no other possible meaning in that context. This necessitated a few changes to the DovahKit backend as well as a change to the INI helper files.

The "trim" helper functions for strings now take a const reference, and return a new string. This is due to the aforementioned restrictions on passing temporaries as non-const references. It's almost certainly wasteful, but an hour of Googling says there's nothing I can do about it.

Several files were missing headers that they needed, with MSVC C++17 letting it slide. C++20 doesn't, so they have been fixed.

u8"" string literals are now char8_t instead of char, so they need to be casted for use in QObject::tr.

It's no longer valid to assign the return value of QAbstractProxyModel::mapToSource to a reference variable.

The dovah::loose_enum helper that we were trialing in ActorBase is now completely broken due to ambiguity caused by the implicit cast-to-enum operators; the built-in (int, int) and (int, enum) operators conflict with each other even though the latter shouldn't even be available for an enum-class. As such, I've dropped the relevant fields on ActorBase down to enum classes, and just set the loader up to load them as uint8_ts with casts where needed -- the low-tech solution.

Fixes unrelated to the C++20 migration are as follows:

Made more of cobb::eight_cc constexpr.

The quest dialog needs to load its working copy before creating its tabs, but needs to do so after creating the "basic data" tab content.

We loaded an extra byte in the NPC_/AIDT loader by mistake. Fixed.

With all of the changes above, the program now builds properly. Scripting hasn't yet been tested, but a brief inspection of form data shows that everything seems to be working exactly as it did before.