Backend fixes and scripting
Fixed a crash in the DEST loader stemming from using the wrong type for the stage count.

The DEST save code now emits a save error if there are too many destruction stages to save.

Instances of our Lua classes now have a __len metamethod that always returns 0, consistent with the behavior of #{}, #{ non_int_key = 5 }, and so on.

Added a global function to the Lua VM: object_is_form, usable for testing whether a userdata is a form.

Quests now treat alias ID -1 as nullptr, and the Lua wrapper for Papyrus properties now properly treats 0xFFFF (i.e. uint16_t(-1)) as -1.

Wrote a test script for viewing a form's top-level Papyrus data.