Papyrus backend work; work on UI scripting
Testing establishes that if a form contains multiple Papyrus scripts with the same name, the last-loaded one overrides the others, clearing their data. Our loader has been updated to mimic this.

If a Lua script attempts to set a Papyrus script-object's name to one used by another Papyrus script-object on the same form, an error will be thrown.