Work on UI scripting: DovahKitLuaCompatibleItem and friends
Started work on a system that will hopefully allow us to manage ranges of Qt model items within Lua wrappers. The key element of this whole system is DovahKitLuaPersistentIndexRange, which is analogous to QPersistentModelIndex but for an entire range of model indices; this in turn requires fully custom model systems for support.

A QComboBox wrapper for Lua will be how we test this whole system.