Work on UI scripting: work on model item support
Lua wrappers for single model items are now widget-independent and do not have a widget pointer. There is a one-to-one correlation between model-observer structs and wrapper instances.

In order to avoid difficulties with the model-observer system, we no longer subclass the observers to add a refcount. Instead, we retain two parallel vectors, one of observer pointers and one of refcounts. We could probably just have a single vector of structs, where instead of subclassing the model-observer the struct just has an observer pointer. Maybe I'll do it next commit.

Added wrapper::teardown, so that we're not just running code inline in the __gc function.

Added dropdown:append_item.