Work on UI scripting
Fixed a bug in the teardown process that could cause a crash if the script was terminated instead of being allowed to stop on its own (i.e. clicking "Stop script" instead of just closing all windows).

The VM core now attempts to keep track of how many script widgets exist. We can use this to limit how many widgets, in total, are allowed to exist, and at some point, we should do precisely that.