Work on UI scripting: important fixes to widget lifetime
Exposed QObject::objectName to scripts as widget.name.

Fixed a very nasty edge-case in our handling of abandoned widget hierarchies: because many widget operations, including widget:add_child, are fire-and-forget and do not block the script, it's possible for a parent widget to go out of scope after we have requested that a child be appended to it but before the child actually is appended to it. This will result in the parent being marked for delete before the child is appended, and then (due to the order in which we process cross-thread tasks) deleted after the child is appended, which in turn deletes the child out from under Lua and leaves Lua with a dangling pointer. We have solved this by allowing a widget that has been newly-appended to a widget hierarchy to rescue that hierarchy from deletion at the last moment.

As part of the above fix, we've also made it so that hierarchies can be linked by a QButtonGroup (i.e. if radio buttons in two different hierarchies belong to the same group), such that a Lua-referenced widget or button group in any hierarchy rescues all linked hierarchies from deletion.