Work on scripting: Papyrus made to-spec; FormPicker fixes
Lua can no longer access Papyrus scripts or properties by index, as these are not (in a "semantic" sense) indexed collections. Under the hood, we implement them as indexed collections just so we can provide Lua with access to form parts in a clear and consistent way, but the correct way to think about the data is as an unordered collection of named elements, so that's how we want script authors to conceptualize it and access it.

Fixed several assertion failures related to scripted FormPickers: when a form is nullptr, we need to explicitly return nil instead of trying to return a wrapper.