Work on form scripting: work on editing Papyrus data
When saving, Papyrus properties now emit normal errors/warnings if they run into problems, instead of failing an assertion.

The dovah::detailed_notice struct can now hold a vector of arbitrary error-code-specific strings. Currently, this is unused, but I hope to use it in the future to allow Papyrus load/save errors to indicate script names, property names, and so on.

Fixes to the Lua APIs for creating and removing Papyrus properties on a script.