Critical fixes to script VM; work on Papyrus API
When the script thread sends the "formModificationImminent" and "formModified" signals for a form, it now does so in lockstep with the main thread, such that the script thread will wait for the signals to be responded to before proceeding. This prevents a number of edge-cases that arise if a form is modified before the "formModificationImminent" signal is sent *and responded to*.

Added papyrus_script:remove_all_properties.

You can now overwrite a Papyrus script-object's properties collection with nil or with the properties collection of another Papyrus script-object. In the latter scenario, the overwritten collection is emptied, and then all content is copied from the overwriting collection.

The script VM's userdata interface now offers a helper function for zombifying all of the element wrappers in a wrapped collection. This is useful when writing functions which clear a collection's contents.