Work on UI scripting: fixes for widget lifetime
Fixed the widget hierarchy finder: it now properly traverses descendants of root widgets, and applies all criteria to the roots themselves; previously, it would only traverse and test direct children, and not test the roots.

We no longer process model reference loss while the VM is being torn down.

The function to mark objects for deletion will no longer fail an assertion if they have already been marked for deletion.

When a model reference is gained, it can now rescue widgets from deletion. When a model reference is lost, it can now condemn widgets to deletion.

We now use a bool to explicitly track when the VM is being torn down.