Work on UI scripting: work on table views
Lua now has basic (but not nearly robust) access to QTableViews. It's theoretically possible to add rows to a table, though I've not tested that yet; it's also possible to set a table's column header texts.

Made it possible for Qt to auto-convert a QVariant<dovah::form_stub*> to a string of the form "[FORM:01234567]EditorID".

Models for Lua-scripted widgets are now created centrally by the VM core. If we ever need to hook model signals, we can now do so.