Work on UI scripting: work on table scripting
Modified ObservableStandardItemModel to make it possible to supply default data for rows or columns, with that data overridden by anything defined on cells. This functionality is meant for use by Lua, but is untested.

Started work on helper functions for getting and setting model data on a per-cell, per-row, or per-column basis.

Lua scripts can now modify an individual cell's text color.