Work on UI scripting: work on table scripting
You can now access a table view's row and column collections.

You can now access a table view's selection, as a table of rows, columns, or cells.

Made helper functions available for Lua APIs to push and pull QColors to and from the Lua stack.

You can now set the default text color for a table row or column, as well as the text color for a cell.

When setting default values for a row or column in a model, we now emit dataChanged, which should cause the changes to display immediately.

Fixed ObservableStandardItemModel::RoleDefaultSet::forCol.

When reading the length of a table row or table column set, we now use a UI read task, which should be thread-safe.

When we send table selection events to Lua, we now properly remap QModelIndexes from the proxy model back to the source.