TES4/INCC; Lua form type constants
The TES4/INCC subrecord has been identified as an interior cell count. AFAIK the game doesn't read it; presumably the CK does, but I'm not sure why. Our code has been modified accordingly: we've added a specific field name, and we now always save the subrecord, taking a count of interior cells present in the saved file.

Fixed a compiler error stemming from errno_t not actually being a standard type of any kind. MSVC used to allow it; clearly it no longer does.

The (form_types) global made available to Lua was, previously, a list of form type constants, wherein a "form type constant" was a form type signature (i.e. a four-CC) stored as an integer. We have replaced these: now, each "form type constant" is a unique userdata, and these userdata have a "signature" member, a "name" member, and an "is" member function that can be used to type-check any value. That means that you can, for example, use (form_types.activator.is(foo)) to test whether (foo) is an Activator form.