Added a dovah.type Lua API
Added a dovah.type function that behaves exactly identically to the built-in type unless it is given a table or a userdata. If it's given a table or a userdata, then it will attempt to access the object's metatable and rawget its __name field, returning the field if its value is a string. This allows Lua scripts to identify the internal types of native-backed userdata values -- in other words, given two userdata that represent totally different things, you can now tell them apart.