Lua: friendlier class names; dovah.object_is
Our class helpers now allow us to specify a "friendly" metatable name for a class, separate from the hidden registry key used to store the metatable.

All Lua wrappers for forms and UI widgets have "friendly" metatable names, e.g. "form" instead of "dovah.classes.form." These names are returned by the dovah.type function.

Added a dovah.object_is(obj, req) function, which accepts a variable of any type and a string. If the passed-in variable is a table or userdata, this function walks its superclass tree and returns true if the variable is an instance of, or derives from, a class with the given name.