Lua color APIs now accept more syntaxes
Ported over a super-duper-simplified version of the "string scanner" from ReachVariantTool. Really, it's less of a port and more of a rebuild. It helps with the task of checking for (and sometimes extracting) basic tokens in a specific order.

Any Lua API which asks for a color should now accept strings of any format accepted by CSS. Previously, these APIs only accepted Lua tables consisting of RGB(A) arrays or with named "r", "g", "b", and "a" fields. Additionally, the helper function that these APIs rely on no longer assumes that the Lua stack offset that it's pulling a color from is an argument index (i.e. we use luaL_error, not luaL_argerror).