Initial implementation of model-observer property handlers
Our wrapper_metatable class is now capable of allowing an "extra class setup" function to run. This can be used to push C-closures into a class's getter or setter lists.

We now have an initial implementation of model observer property handlers (MOPHs). Each handler consists of a name, to be made available to Lua, and two function pointers for transferring a value between a QVariant and the Lua stack. There are also helpers for creating [gs]etters that rely on these two functions, using C-closures.

The Lua class for table view cells now defines a MOPH list and uses the aforementioned helper functions to grant access to its text (Qt::DisplayRole) and text color (Qt::ForegroundRole).

Next steps include: making it possible for MOPH value-transfer functions to report errors in a Lua value, so that getters can throw them outright and other functions can maybe log them as warnings; making the MOPH "getter," "setter," "push_getter," and "push_getter" functions take a class metatable key as an upvalue/argument, so we don't need to template them and can move them out of the header; and making it possible for an MOPH to "transform" (i.e. modify) a property instead of overwriting it (e.g. an MOPH for text alignment should be able to handle "center unchanged" based on the pre-existing value).