Work on MOPHs: work on transforming existing values
It is now possible for an MOPH to transform an existing value when setting, instead of overwriting the existing value entirely. We currently use this when changing table view cells' text alignment, to allow the use of an "unchanged" keyword for either or both axes. We'll also be able to use it for font properties (e.g. italics, bold, etc.) in the future.

Added an "alignment" property for table view cells.