Work on MOPHs: error reporting; MOPH system is now feature-complete
MOPH "pull" functions can now use luaL_error and friends to report errors. If the "pull" function is invoked as part of a model observer setter, then the error will be thrown verbatim; if the function is invoked as part of handler_set::extract, then the error text will be emitted as a warning.

The MOPH list for table_view_cell is now a static class member and therefore accessible to other C++ code.

The table_view:append_row API now allows you to specify table arguments, which will be interpreted by the table_view_cell MOPH list. With this, MOPHs as a system are now feature-complete.