Work on UI scripting: fixes and polishing
Fixed a bug that prevented Lua scripts from properly gaining a reference to a table view column or row if the first cell in the column or row was missing.

If the user attempts to close the script window while a script is running, a confirmation dialog will now appear.